Refactors data types for OLED functions
Updates the `delay_ms` function and OLED-related functions to use `uint32_t` and `uint8_t` types, respectively.
This change ensures type consistency and compatibility across different architectures, improving code reliability and portability.
Also reorders source entries in .cproject.

+++ Drivers/OLED_Software_I2C/oled.c
@@ -13,7 +13,7 @@
 //[6]0 1 2 3 ... 127	
 //[7]0 1 2 3 ... 127
 
-void delay_ms(unsigned long ms) 
+void delay_ms(uint32_t ms)
 {
     mspm0_delay_ms(ms);
 }
